requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
passlib[bcrypt]
python-dotenv
pydantic[email]
